Product CRUD v3.0

+++ public/js/ajaxscript.js
@@ -26,27 +26,64 @@
         $('#frmProducts').trigger("reset");
         $('#myModal').modal('show');
     });
-    //delete product and remove it from list
+
+    // delete product and remove it from list
     $(document).on('click','.delete-product',function(){
         var product_id = $(this).val();
-         $.ajaxSetup({
-            headers: {
-                'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')
-            }
-        })
-        $.ajax({
-            type: "DELETE",
-            url: url + '/' + product_id,
-            success: function (data) {
-                console.log(data);
-                $("#product" + product_id).remove();
-            },
-            error: function (data) {
-                console.log('Error:', data);
-            }
 
+        $.confirm({
+            text: "Are you sure you want to delete this product? <strong>This action cannot be undone.</strong>",
+            title: "Confirmation required",
+            confirm: function(button) {
+              
+            $.ajaxSetup({
+                headers: {
+                    'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')
+                }
+            })
+            $.ajax({
+                type: "DELETE",
+                url: url + '/' + product_id,
+                success: function (data) {
+                    console.log(data);
+                    $("#product" + product_id).remove();
+                },
+                error: function (data) {
+                    console.log('Error:', data);
+                }
+    
+            });
+            alert("Product successfully deleted");
+            },
+            cancel: function(button) {
+            // nothing to do
+            },
+            confirmButton: "Yes I am sure",
+            confirmButtonClass: "btn-danger",
+            cancelButtonClass: "btn-default",
         });
+        
     });
+    // $(document).on('click','.delete-product',function(){
+    //     var product_id = $(this).val();
+    //      $.ajaxSetup({
+    //         headers: {
+    //             'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')
+    //         }
+    //     })
+    //     $.ajax({
+    //         type: "DELETE",
+    //         url: url + '/' + product_id,
+    //         success: function (data) {
+    //             console.log(data);
+    //             $("#product" + product_id).remove();
+    //         },
+    //         error: function (data) {
+    //             console.log('Error:', data);
+    //         }
+
+    //     });
+    // });
     //create new product / update existing product
     $("#btn-save").click(function (e) {
         $.ajaxSetup({
